technoatom-2020-1/lection10/python_all/requirements.txt
pytest
flask
requests
allure-pytest
pytest-xdist
pymysql
sqlalchemy
faker